backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.28
aiosqlite>=0.20.0
pydantic>=2.6.4
pydantic-settings>=2.2.1
openai>=1.14.0
twilio>=9.0.0
pytest>=8.1.1
httpx>=0.27.0
pytest-asyncio>=0.23.6
python-dotenv>=1.0.1
python-docx>=1.1.0
faiss-cpu>=1.7.4
numpy>=1.26.0
